pyproject.toml
[project]
name = "trabalho-iii-persistencia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beanie>=2.0.1",
    "fastapi>=0.128.0",
    "motor>=3.7.1",
    "odmantic>=1.0.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
